move get paged events to events
